RDBC-959: move commands internally into operation files
