+++ pom.xml
@@ -115,6 +115,18 @@
             <artifactId>jackson-databind</artifactId>
             <version>2.16.2</version>
         </dependency>
+        <dependency>
+            <groupId>jakarta.validation</groupId>
+            <artifactId>jakarta.validation-api</artifactId>
+            <version>3.0.2</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.hibernate.validator</groupId>
+            <artifactId>hibernate-validator</artifactId>
+            <version>8.0.1.Final</version>
+        </dependency>
+
 
 
         <!-- test dependencies -->
