+++ src/main/java/net/ravendb/client/documents/AI/TextPart.java
@@ -3,9 +3,11 @@ package net.ravendb.client.documents.AI;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.fasterxml.jackson.databind.node.JsonNodeFactory;
+import jakarta.validation.constraints.NotBlank;
 
 public final class TextPart extends ContentPart {
 
+    @NotBlank(message = "Text cannot be null or empty")
     @JsonProperty("text")
     private String text;
